Database design for onchain + ethflow orders (#499)
This PR contains the db design for the generic on-chain orders + ethflow orders.

The generic on-chain orders design contains all the information that is generic to any on-chain placed order. The data is retrieved from this event
Specifically:

address indexed sender, -> stored in generic db
GPv2Order.Data order, -> all information will be stored in the standard order database
OnchainSignature signature, -> all information will be stored in the standard order database
bytes data -> this is custom and depends on the order type. E.g. for the eth-flow orders, it contains the user_valid_to and quote_id.
Additionally, we will also parse the quote_id from the bytes data field, but this information does not need to be stored.

The only information that is additionally stored for eth-flow orders is the user_valid_to.
All the information will later one be linked via joins of the different tables on the order_uid - their primary key.

In order to see how the is_reorged field is used, please visit: #500

Test Plan
CI