[Trivial] Make swagger connect to production by default (#516)
This PR flips the order in which our different endpoints are defined in the swagger document to make hitting the prod API the default. We have shared the API docs with a bunch of external developers and solvers and the most common mistake is them not pointing at the correct endpoint (therefore searching a tx hash or uid that happened in prod and is not found).

While for internal debugging it might be nice to be by default connected to staging, I feel like the number of people we need to explain this to internally is fewer than the number of external people using this service.