General onchain order parsing logic (#536)
Implements the general parsing of onchain orders.
PR should be reviewed with one follow-up PRs: 
- custom trait implementation for ethflow orders (https://github.com/cowprotocol/services/pull/537)

This PR handles the general onchain order parsing for onchain orders.  It implements all the parsing logic that is common for all the orders and defines a trait `OnchainOrderParsing` it is calling into for the custom parsing. 
In the follow-up PR, we implement the custom OnchainOrderParsing trait for the ethflow orders.

### Test Plan
Unit test only