Pre_interactions database scheme & database operations (#620)
The ethflow needs to have a pre_interaction that wraps all the eth in the ethflow contract. We decided that we want to go with a more general pre-interaction approach.

This PR provides the db scheme and operations for the pre-interactions: inserting and reading of the interactions + reading the interactions for all orders.