Encode pre-interactions for settlement (#623)
Encodes the pre_interactions of the orders into the settlement. It also filters for duplicated interactions and removes them.

Additionally, it lifts the interaction trait from the solver crate into the shared crate: We want to use the same trait Interaction with the same encode interface, as for pre-interactions of orders.