Improve Handling of Slippage for AMMs (#659)
In #589 et. al., we introduced a new mechanism for computing slippage values. Specifically, we now have a hybrid relative/absolute slippage approach, where we apply a relative slippage capped at a maximum absolute value denominated in the network's native token.

This works well for DEX aggregators since the slippage calculation relies on external prices, and DEX aggregators **always** apply slippage to a token amount for a traded token (which are guaranteed to have external prices).

However, this does not work well for solvers that use the "liquidity" abstraction. Specifically, we fetch liquidity for all traded tokens **AND** a set of base tokens. While the traded tokens all have external prices, base tokens are not guaranteed to have external prices. This means that it is possible to try and compute slippage for a token amount for which there is no price information.

This PR changes the capped slippage computation to:
- First, try and compute the `input_max` token amount with slippage added normally (that is, use the token amount and external token price to compute the actual slippage amount to use)
- Then, try and compute the relative slippage amount using the `output` token amount, and apply this relative slippage to the input token amount. This means that for cases where there is no external price for the input token, but there is an external price for the output token, we still get a correct capped slippage amount
- Finally, just use the configured relative slippage amount without a cap. Ideally, we would be able to use a combination of external prices, clearing prices, executed AMM amounts, in order to compute a capped slippage amount, but that is a **much** more complicated and should rarely happen.

### Test Plan

Added a unit test. Existing tests continue to pass.

### Release notes

We can re-enable absolute slippage caps for **all** solvers (right now, they are only enabled for DEX aggregator solvers).