Fix HTTP Solver Liquidity Indexing (#751)
I was looking into why Quasimodo was not being as successful as it should in the high volume times this evening, and noticed in local tests that the index of the liquidity AMM models didn't line up with the liquidity in the `SettlementContext` used for settlement encoding.

The reason is that the index that we use for matching AMM models to liquidity when encoding was being generated by enumerating all liquidity _with limit orders (like 0x) filtered out_ while the liquidity vector in the `SettlementContext` _still contained the limit orders_.

This PR addresses this by splitting the `orders` and `amms` early on in the `prepare_model` function to make sure we use them consistently throughout. This might also be the cause for Uniswap V3 liquidity not encoding correctly, but I did not confirm this.

### Test Plan

Added a unit test.